+++ Platform.SDK.Components/BibleComponents/BibleReader.razor.cs
@@ -1,4 +1,8 @@
 #region usings
+
+using System;
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
 using Platform.API.Models;
 using Platform.SDK.Services;